refactoring data module to domain, adding bank context.

+++ src/roles.rs
@@ -1,9 +1,9 @@
 pub trait CheckingAccount {
-    fn increment(&mut self, amount: f32) -> f32;
+    fn increment(&mut self, amount: f64) -> f64;
 }
 
 pub trait SavingsAccount {
-    fn decrement(&mut self, amount: f32) -> f32;
+    fn decrement(&mut self, amount: f64) -> f64;
 }
 
 pub trait AccountLogger {
